Favorites component moved to components and counter removed.
