+++ client/src/components/favorites/index.js
