+++ client/src/components/favorites/style.css
